// src/app/models/tracker.model.ts
export class Tracker {
  id: string;
  firstName: string;
  lastName: string;
  mobile: number;
  otherNumber: number;
  city: string;
  state: string;
  isCovidPositiveRecovered: string;
  isReadyDonatePlasma: string;
  isDonatedForPast30Days: string;
  bloodGroup: string;
  createdDate = new Date();
  updateDate: Date;
  createdBy: string;
  updatedBy: string;

  // constructor() {
  //   this.firstName = 'tste';
  //   this.lastName = 'tt';
  //   this.mobile = 1231231231;
  //   this.otherNumber = 1233123123;
  //   this.city = 'tt';
  //   this.state = 'Tamil Nadu';
  //   this.isCovidPositiveRecovered = 'Yes';
  //   this.isReadyDonatePlasma = 'Yes';
  //   this.isDonatedForPast30Days = 'Yes';
  //   this.bloodGroup = 'A+';
  //   this.createdDate = new Date();
  //   this.updateDate = new Date();
  //   this.createdBy = 'tt';
  //   this.updatedBy = 'tt';
  // }
}

// export enum BloodGroup {
//   ['A+'] = 'A+',
//   ['A-'] = 'A-',
//   ['B+'] = 'B+',
//   ['B-'] = 'B-',
//   ['O+'] = 'B-',
//   ['O-'] = 'B-',
//   ['AB+'] = 'AB+',
//   ['AB-'] = 'AB-'
// }